+++ lib/chewy/type/actions.rb
@@ -23,11 +23,19 @@ module Chewy
         #   UsersIndex::User.sync
         #
         # @see Chewy::Type::Syncer
+        # @param parallel [true, Integer, Hash] options for parallel execution or the number of processes
         # @return [Hash{Symbol, Object}, nil] a number of missing and outdated documents reindexed and their ids, nil in case of errors
-        def sync
-          syncer = Syncer.new(self)
+        def sync(parallel: nil)
+          syncer = Syncer.new(self, parallel: parallel)
           count = syncer.perform
           {count: count, missing: syncer.missing_ids, outdated: syncer.outdated_ids} if count
+        end
+
+        # A {Chewy::Journal} instance for the particular type
+        #
+        # @return [Chewy::Journal] journal instance
+        def journal
+          @journal ||= Chewy::Journal.new(self)
         end
       end
     end