+++ spec/chewy/rake_helper_spec.rb
@@ -21,8 +21,16 @@ describe Chewy::RakeHelper, :orm do
 
   let!(:cities) { Array.new(3) { |i| City.create!(name: "Name#{i + 1}") } }
   let!(:countries) { Array.new(2) { |i| Country.create!(name: "Name#{i + 1}") } }
+  let(:journal) do
+    Chewy::Stash::Journal.import([
+      {index_name: 'places', type_name: 'city', action: 'index', references: cities.first(2).map(&:id).map(&:to_s).map(&:to_json), created_at: 2.minutes.since},
+      {index_name: 'places', type_name: 'country', action: 'index', references: [countries.first.id.to_s.to_json], created_at: 4.minutes.since}
+    ])
+  end
 
   describe '.reset' do
+    before { journal }
+
     specify do
       output = StringIO.new
       expect { described_class.reset(output: output) }
@@ -31,6 +39,9 @@ describe Chewy::RakeHelper, :orm do
 \\AResetting PlacesIndex
   Imported PlacesIndex::City for \\d+s, stats: index 3
   Imported PlacesIndex::Country for \\d+s, stats: index 2
+  Applying journal to \\[PlacesIndex::City, PlacesIndex::Country\\], 3 entries, stage 1
+  Imported PlacesIndex::City for \\d+s, stats: index 2
+  Imported PlacesIndex::Country for \\d+s, stats: index 1
   Imported Chewy::Stash::Specification for \\d+s, stats: index 1
 Resetting UsersIndex
   Imported Chewy::Stash::Specification for \\d+s, stats: index 1
@@ -46,6 +57,9 @@ Total: \\d+s\\Z
 \\AResetting PlacesIndex
   Imported PlacesIndex::City for \\d+s, stats: index 3
   Imported PlacesIndex::Country for \\d+s, stats: index 2
+  Applying journal to \\[PlacesIndex::City, PlacesIndex::Country\\], 3 entries, stage 1
+  Imported PlacesIndex::City for \\d+s, stats: index 2
+  Imported PlacesIndex::Country for \\d+s, stats: index 1
   Imported Chewy::Stash::Specification for \\d+s, stats: index 1
 Total: \\d+s\\Z
       OUTPUT
@@ -242,6 +256,113 @@ Total: \\d+s\\Z
 Total: \\d+s\\Z
         OUTPUT
       end
+    end
+  end
+
+  describe '.journal_apply' do
+    specify { expect { described_class.journal_apply }.to raise_error ArgumentError }
+    specify do
+      output = StringIO.new
+      described_class.journal_apply(time: Time.now, output: output)
+      expect(output.string).to match(Regexp.new(<<-OUTPUT, Regexp::MULTILINE))
+\\AApplying journal entries created after [+-:\\d\\s]+
+No journal entries were created after the specified time
+Total: \\d+s\\Z
+      OUTPUT
+    end
+
+    context do
+      before { journal }
+
+      specify do
+        output = StringIO.new
+        expect { described_class.journal_apply(time: Time.now, output: output) }
+          .to update_index(PlacesIndex::City)
+        expect(output.string).to match(Regexp.new(<<-OUTPUT, Regexp::MULTILINE))
+\\AApplying journal entries created after [+-:\\d\\s]+
+  Applying journal to \\[PlacesIndex::City, PlacesIndex::Country\\], 3 entries, stage 1
+  Imported PlacesIndex::City for \\d+s, stats: index 2
+  Imported PlacesIndex::Country for \\d+s, stats: index 1
+Total: \\d+s\\Z
+        OUTPUT
+      end
+
+      specify do
+        output = StringIO.new
+        expect { described_class.journal_apply(time: 3.minutes.since, output: output) }
+          .not_to update_index(PlacesIndex::City)
+        expect(output.string).to match(Regexp.new(<<-OUTPUT, Regexp::MULTILINE))
+\\AApplying journal entries created after [+-:\\d\\s]+
+  Applying journal to \\[PlacesIndex::Country\\], 1 entries, stage 1
+  Imported PlacesIndex::Country for \\d+s, stats: index 1
+Total: \\d+s\\Z
+        OUTPUT
+      end
+
+      specify do
+        output = StringIO.new
+        expect { described_class.journal_apply(time: Time.now, only: PlacesIndex::City, output: output) }
+          .to update_index(PlacesIndex::City)
+        expect(output.string).to match(Regexp.new(<<-OUTPUT, Regexp::MULTILINE))
+\\AApplying journal entries created after [+-:\\d\\s]+
+  Applying journal to \\[PlacesIndex::City\\], 2 entries, stage 1
+  Imported PlacesIndex::City for \\d+s, stats: index 2
+Total: \\d+s\\Z
+        OUTPUT
+      end
+
+      specify do
+        output = StringIO.new
+        expect { described_class.journal_apply(time: Time.now, except: PlacesIndex::City, output: output) }
+          .not_to update_index(PlacesIndex::City)
+        expect(output.string).to match(Regexp.new(<<-OUTPUT, Regexp::MULTILINE))
+\\AApplying journal entries created after [+-:\\d\\s]+
+  Applying journal to \\[PlacesIndex::Country\\], 1 entries, stage 1
+  Imported PlacesIndex::Country for \\d+s, stats: index 1
+Total: \\d+s\\Z
+        OUTPUT
+      end
+    end
+  end
+
+  describe '.journal_clean' do
+    before { journal }
+
+    specify do
+      output = StringIO.new
+      described_class.journal_clean(output: output)
+      expect(output.string).to match(Regexp.new(<<-OUTPUT, Regexp::MULTILINE))
+\\ACleaned up 2 journal entries
+Total: \\d+s\\Z
+      OUTPUT
+    end
+
+    specify do
+      output = StringIO.new
+      described_class.journal_clean(time: 3.minutes.since, output: output)
+      expect(output.string).to match(Regexp.new(<<-OUTPUT, Regexp::MULTILINE))
+\\ACleaning journal entries created before [+-:\\d\\s]+
+Cleaned up 1 journal entries
+Total: \\d+s\\Z
+      OUTPUT
+    end
+
+    specify do
+      output = StringIO.new
+      described_class.journal_clean(only: PlacesIndex::City, output: output)
+      expect(output.string).to match(Regexp.new(<<-OUTPUT, Regexp::MULTILINE))
+\\ACleaned up 1 journal entries
+Total: \\d+s\\Z
+      OUTPUT
+    end
+
+    specify do
+      output = StringIO.new
+      described_class.journal_clean(except: PlacesIndex::City, output: output)
+      expect(output.string).to match(Regexp.new(<<-OUTPUT, Regexp::MULTILINE))
+\\ACleaned up 1 journal entries
+Total: \\d+s\\Z
+      OUTPUT
     end
   end
 end